SMMS-SchoolMeal-AI-local/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
faiss-cpu
sentence-transformers
networkx
sqlalchemy
pyodbc
python-dotenv
scikit-learn
